rename PaymentViewModel to PaymentWrapper
